requirements.txt
Flask==3.1.0
Flask-MySQLdb==2.0.0
PyJWT==2.9.0
pytest==8.3.4